refactor: reorganize CLI commands and update import paths for generators
